move no-op Compressor into message_compress.h/.cc
this is so that we can later have a number of Compressors be automatically
registered in the CompressorRegistry singleton.

As it is right now, there would be a circular import that seems like it would
be an issue (although I could be wrong) - where message_compress.h has to
import null_compressor.h and null_compressor.h has to import
message_compress.h.

I'll soon add other Compressor implementations such as one for gzip/deflate
alongside NullCompressor.